Update LOTE.css
hey im not a coder idk what any of this means but there was a bug pointed on roll20 forums about the watershaper character sheets. line 27-28.  ive got a game coming up and wondering how long it'd take to fix this bug on roll20? most of my players want to be waterbenders :(